app/requirements.txt
# Core Flask dependencies
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2

# Machine Learning - Use CPU-only TensorFlow for better compatibility
tensorflow-cpu==2.13.1
numpy==1.24.3
pandas==2.0.3

# Use pre-compiled scikit-learn to avoid compilation issues
scikit-learn==1.3.2

# Optional visualization libraries (commented out to reduce build time)
# matplotlib==3.7.2
# seaborn==0.12.2
# plotly==5.15.0

# Production server
gunicorn==21.2.0
